optimuspy-wavefront_subpath/requirements/prod.txt
django
django-bootstrap-v5
crispy-bootstrap5
django-crispy-forms
social-auth-app-django
python-decouple
Celery
django-extensions
bokeh
psycopg2
uvicorn [standard]
channels
channels-rabbitmq
py-cpuinfo
django-celery-results
